% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbBinom.R
\name{FuncBin}
\alias{FuncBin}
\title{Probabilidad Binomial}
\usage{
FuncBin(y, p = 0.5)
}
\arguments{
\item{y}{indica la variable de la distribución binomial}

\item{p}{indica la probabilidad}
}
\value{
Devuelve una lista con el coeficiente, la probabilidad binomial, la probabilidad acumulativa y las graficas correspondientes a las probabilidades
}
\description{
Describe el número de éxitos al realizar n experimentos independientes entre sí, acerca de una variable aleatoria
}
\examples{
\dontrun{
# Directorio
directorio <- "C:\\\\Users\\\\CICEM\\\\OneDrive\\\\Escritorio\\\\ProbabilidadBinomial\\\\Prueba\\\\Datos.csv"
datos <- read.csv(directorio)
# Imprimimos las primeras filas
head(datos)
Elegimos los datos de la primera columna
y <-  datos[, 1]
y
Elegimos una probabilidad de éxito del 0.5
p = 0.5
Importamos la librería
library(ProbBinom)
Probamos la función
resultado <- FuncBin(y, p)
resultado
}
}
